Crate bdk_file_store

BDK File Store

⚠ bdk_file_store is a development/testing database. It does not natively support backwards compatible BDK version upgrades so should not be used in production.

This is a simple append-only flat file database for persisting bdk_chain changesets.

The main structure is Store which works with any bdk_chain based changesets to persist data into a flat file.

Structs

AggregateChangesetsError

Error type for Store::aggregate_changesets.

EntryIter

Iterator over entries in a file store.

Store

Persists an append-only list of changesets (C) to a single file.

Enums

FileError

Error that occurs due to problems encountered with the file.

IterError

Error type for EntryIter.